refactor: CommonType으로 move
