+++ src/common/type.ts
@@ -1,0 +1,8 @@
+export enum Part {
+  iOS = 'iOS',
+  PLAN = 'PLAN',
+  DESIGN = 'DESIGN',
+  SERVER = 'SERVER',
+  ANDROID = 'ANDROID',
+  WEB = 'WEB',
+}
